# openapi_memgpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MemGPT API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agents/config": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Config",
        "description": "Retrieve the configuration for a specific agent.\n\nThis endpoint fetches the configuration details for a given agent, identified by the user and agent IDs.",
        "operationId": "get_agent_config_api_agents_config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "agent_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}/rename": {
      "patch": {
        "tags": [
          "agents"
        ],
        "summary": "Update Agent Name",
        "description": "Updates the name of a specific agent.\n\nThis changes the name of the agent in the database but does NOT edit the agent's persona.",
        "operationId": "update_agent_name_api_agents__agent_id__rename_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/delete": {
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "Delete Agent",
        "description": "Delete an agent.",
        "operationId": "delete_agent_api_agents_delete_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentRenameRequest": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name",
            "description": "New name for the agent."
          }
        },
        "type": "object",
        "required": [
          "agent_name"
        ],
        "title": "AgentRenameRequest"
      },
      "AgentStateModel": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "The unique identifier of the agent."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the agent."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "The description of the agent."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "The unique identifier of the user associated with the agent."
          },
          "created_at": {
            "type": "integer",
            "title": "Created At",
            "description": "The unix timestamp of when the agent was created."
          },
          "preset": {
            "type": "string",
            "title": "Preset",
            "description": "The preset used by the agent."
          },
          "persona": {
            "type": "string",
            "title": "Persona",
            "description": "The persona used by the agent."
          },
          "human": {
            "type": "string",
            "title": "Human",
            "description": "The human used by the agent."
          },
          "functions_schema": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Functions Schema",
            "description": "The functions schema used by the agent."
          },
          "llm_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LLMConfigModel"
              }
            ],
            "description": "The LLM configuration used by the agent."
          },
          "embedding_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmbeddingConfigModel"
              }
            ],
            "description": "The embedding configuration used by the agent."
          },
          "state": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "The state of the agent."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "user_id",
          "created_at",
          "preset",
          "persona",
          "human",
          "functions_schema",
          "llm_config",
          "embedding_config"
        ],
        "title": "AgentStateModel"
      },
...
      "DeleteAgentRequest": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id",
            "description": "The optional ID the agent."
          },
          "agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Name",
            "description": "The optional name of the agent."
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "agent_name"
        ],
        "title": "DeleteAgentRequest"
      },
...
      "EmbeddingConfigModel": {
        "properties": {
          "embedding_endpoint_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Endpoint Type",
            "default": "openai"
          },
          "embedding_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Endpoint",
            "default": "https://api.openai.com/v1"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model",
            "default": "text-embedding-ada-002"
          },
          "embedding_dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Dim",
            "default": 1536
          },
          "embedding_chunk_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Chunk Size",
            "default": 300
          }
        },
        "type": "object",
        "title": "EmbeddingConfigModel"
      },
...
      "GetAgentResponse": {
        "properties": {
          "agent_state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AgentStateModel"
              }
            ],
            "description": "The state of the agent."
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "The list of data sources associated with the agent."
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At",
            "description": "The unix timestamp of when the agent was last run."
          }
        },
        "type": "object",
        "required": [
          "agent_state",
          "sources"
        ],
        "title": "GetAgentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMConfigModel": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "default": "gpt-4"
          },
          "model_endpoint_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Endpoint Type",
            "default": "openai"
          },
          "model_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Endpoint",
            "default": "https://api.openai.com/v1"
          },
          "model_wrapper": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Wrapper"
          },
          "context_window": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          }
        },
        "type": "object",
        "title": "LLMConfigModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://memgpt.localhost"
    },
    {
      "url": "http://localhost:8283"
    },
    {
      "url": "http://localhost:8083"
    }
  ]
}